Remove unnecessary cast.

+++ jdcp-core/src/main/java/ca/eandb/jdcp/job/CompositeParallelizableJob.java
@@ -95,7 +95,7 @@ public final class CompositeParallelizableJob implements ParallelizableJob {
       if (jobProgress == null) {
         jobProgress = new double[jobs.size()];
       }
-      totalProgress += (progress - jobProgress[jobNumber]) / (double) jobs.size();
+      totalProgress += (progress - jobProgress[jobNumber]) / jobs.size();
       jobProgress[jobNumber] = progress;
       return true;
     }
